requirements.txt
moviepy==2.1.1
gTTS==2.5.4
requests==2.32.0
python-dotenv==0.21.0
dropbox==12.0.2
selenium==4.27.1
undetected-chromedriver==3.5.5
beautifulsoup4==4.12.3
schedule==1.2.2
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
transformers==4.45.2
deep-translator==1.11.4
